Preserve occupied Kuma groups during races
