+++ test/shared/uptime-kuma/client/protocol.test.ts
@@ -1,101 +1,10 @@
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
-import { stub } from "@std/testing/mock";
 import { FakeTime } from "@std/testing/time";
-import {
-  getSubrequestUsage,
-  runWithSubrequestBudget,
-  withSubrequestAllowance,
-} from "#shared/subrequest-budget.ts";
-import {
-  createUptimeKumaClient,
-  type UptimeKumaSocket,
-  uptimeKumaClientApi,
-  uptimeKumaSocketFactory,
-} from "#shared/uptime-kuma/client.ts";
-import type { UptimeKumaConfig } from "#shared/uptime-kuma/config.ts";
+import { createUptimeKumaClient } from "#shared/uptime-kuma/client.ts";
+import { FakeSocket } from "./support.test.ts";
 
-type Listener = (...args: unknown[]) => void;
-type Reply = unknown | ((...args: unknown[]) => unknown);
-
-class FakeSocket implements UptimeKumaSocket {
-  calls: Array<{ args: unknown[]; event: string }> = [];
-  connectError: unknown = null;
-  disconnected = false;
-  timeoutMs = 0;
-  #listeners = new Map<string, Set<Listener>>();
-  #replies = new Map<string, Reply>();
-
-  connect(): void {
-    this.emitEvent(
-      this.connectError === null ? "connect" : "connect_error",
-      this.connectError,
-    );
-  }
-
-  disconnect(): void {
-    this.disconnected = true;
-  }
-
-  emitEvent(event: string, ...args: unknown[]): void {
-    for (const listener of this.#listeners.get(event) ?? []) {
-      listener(...args);
-    }
-    this.#listeners.delete(event);
-  }
-
-  emitWithAck(event: string, ...args: unknown[]): Promise<unknown> {
-    this.calls.push({ args, event });
-    if (!this.#replies.has(event)) {
-      return Promise.reject(new Error(`No reply for ${event}`));
-    }
-    const reply = this.#replies.get(event);
-    return Promise.resolve(
-      typeof reply === "function" ? reply(...args) : reply,
-    );
-  }
-
-  off(event: string, listener: Listener): void {
-    this.#listeners.get(event)?.delete(listener);
-  }
-
-  listenerCount(event: string): number {
-    return this.#listeners.get(event)?.size ?? 0;
-  }
-
-  once(event: string, listener: Listener): void {
-    const listeners = this.#listeners.get(event) ?? new Set<Listener>();
-    listeners.add(listener);
-    this.#listeners.set(event, listeners);
-  }
-
-  reply(event: string, value: Reply): void {
-    this.#replies.set(event, value);
-  }
-
-  timeout(milliseconds: number): UptimeKumaSocket {
-    this.timeoutMs = milliseconds;
-    return this;
-  }
-}
-
-const config: UptimeKumaConfig = {
-  intervalSeconds: 900,
-  password: "password",
-  url: "https://kuma.example.test/status",
-  username: "owner",
-};
-
-const returnSocket = (socket: UptimeKumaSocket) =>
-  stub(uptimeKumaSocketFactory, "create", () => Promise.resolve(socket));
-
-const expectDisconnected = (socket: FakeSocket): void => {
-  expect(socket.disconnected).toBe(true);
-  expect(socket.listenerCount("connect")).toBe(0);
-  expect(socket.listenerCount("connect_error")).toBe(0);
-};
-
-describe("Uptime Kuma Socket.IO client", () => {
+describe("Uptime Kuma Socket.IO protocol", () => {
   test("logs in with the username and password payload", async () => {
     const socket = new FakeSocket();
     socket.reply("login", { ok: true, token: "session-token" });
@@ -328,85 +237,6 @@ describe("Uptime Kuma Socket.IO client", () => {
 
     createUptimeKumaClient(socket).disconnect();
 
-    expectDisconnected(socket);
-  });
-
-  test("connects before returning a client", async () => {
-    const socket = new FakeSocket();
-    using _factory = returnSocket(socket);
-
-    const client = await uptimeKumaClientApi.connect(config);
-    client.disconnect();
-
-    expectDisconnected(socket);
-  });
-
-  test("disconnects after a connection error", async () => {
-    const socket = new FakeSocket();
-    socket.connectError = new Error("socket refused");
-    using _factory = returnSocket(socket);
-
-    await expect(uptimeKumaClientApi.connect(config)).rejects.toThrow(
-      "socket refused",
-    );
-    expectDisconnected(socket);
-  });
-
-  test("uses a safe message for a malformed connection error", async () => {
-    const socket = new FakeSocket();
-    socket.connectError = { password: "must not leak" };
-    using _factory = returnSocket(socket);
-
-    await expect(uptimeKumaClientApi.connect(config)).rejects.toThrow(
-      "Uptime Kuma connection failed.",
-    );
-  });
-
-  test("places the Socket.IO endpoint under the configured base path", async () => {
-    await runWithSubrequestBudget(async () => {
-      const socket = await uptimeKumaSocketFactory.create(config);
-      const manager = socket as unknown as {
-        io: {
-          opts: {
-            autoConnect: boolean;
-            path: string;
-            reconnection: boolean;
-            timeout: number;
-            transports: unknown[];
-          };
-        };
-      };
-
-      expect(manager.io.opts).toMatchObject({
-        autoConnect: false,
-        path: "/status/socket.io",
-        reconnection: false,
-        timeout: 10_000,
-        transports: ["websocket"],
-      });
-      expect(getSubrequestUsage().external).toBe(1);
-      socket.disconnect();
-    });
-  });
-
-  test("places a root Kuma server at the root Socket.IO path", async () => {
-    const socket = await uptimeKumaSocketFactory.create({
-      ...config,
-      url: "https://kuma.example.test",
-    });
-    const manager = socket as unknown as { io: { opts: { path: string } } };
-
-    expect(manager.io.opts.path).toBe("/socket.io");
-    socket.disconnect();
-  });
-
-  test("names a blocked Kuma connection in the subrequest error", async () => {
-    await expect(
-      withSubrequestAllowance({ database: 50, external: 0, total: 50 }, () =>
-        uptimeKumaSocketFactory.create(config),
-      ),
-    ).rejects.toThrow(
-      "Blocked external operation: Uptime Kuma socket connection",
-    );
+    expect(socket.disconnected).toBe(true);
   });
 });
